Extract redundant trait functions
