+++ crates/sncast/src/starknet_commands/mod.rs
@@ -7,4 +7,4 @@ pub mod multicall;
 pub mod script;
 pub mod show_config;
 pub mod tx_status;
-pub mod verification;
+pub mod verify;
