--- crates/sncast/src/starknet_commands/verification/mod.rs
@@ -1,4 +1,0 @@
-pub mod base;
-pub mod verify;
-mod voyager;
-mod walnut;
