+++ crates/sncast/tests/e2e/mod.rs
@@ -8,4 +8,4 @@ mod multicall;
 mod script;
 mod show_config;
 mod tx_status;
-mod verification;
+mod verify;
