+++ crates/sncast/tests/e2e/verify/helpers.rs
