+++ crates/sncast/tests/e2e/verify/walnut.rs
